requirements.txt
Django==5.0.1
djangorestframework==3.14.0
python-dotenv==1.0.0
pillow==10.2.0
psycopg2-binary==2.9.9
gunicorn==21.2.0
drf-yasg==1.21.7
django-cors-headers==4.3.1
celery==5.3.6
redis==5.0.1
